fix(proto): correct OpaqueAuth deserialization for RFC 5531 compliance
Root cause of "bytes remaining on stream" / deserialization failures
when Linux NFS client (mount.nfs4) connects to nextnfs:

1. OpaqueAuth was deserialized as an XDR discriminated union via
   serde derive, but RFC 5531 defines opaque_auth as a struct with
   flavor (u32) + opaque body (u32 length + data + padding). The
   derived enum deserializer skipped the body length prefix, causing
   all subsequent fields to be misaligned by 4+ bytes.

2. AuthUnix.stamp was u64 but RFC 5531 authsys_parms defines it as
   unsigned int (u32). This consumed 4 extra bytes per credential.

Fix: Replace derived Serialize/Deserialize with custom implementations
that properly handle the opaque body wrapper. The deserializer reads
flavor + opaque bytes, then parses the body based on flavor. The
serializer writes flavor + body length + serialized body bytes.

+++ proto/src/rpc_proto.rs
@@ -13,14 +13,14 @@ use super::{
 
 #[derive(Clone, Debug, Deserialize, Serialize, Default)]
 pub struct AuthUnix {
-    pub stamp: u64,
+    pub stamp: u32,
     pub machinename: String,
     pub uid: u32,
     pub gid: u32,
     pub gids: Vec<u32>,
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug)]
 #[repr(u32)]
 pub enum OpaqueAuth {
     AuthNull(Vec<u8>) = 0,
